Added logging and javadoc

+++ trunk/src/main/java/org/unitils/dbmaintainer/clean/DefaultDBCleaner.java
@@ -16,6 +16,7 @@
 package org.unitils.dbmaintainer.clean;
 
 import org.apache.commons.configuration.Configuration;
+import org.apache.log4j.Logger;
 import org.unitils.core.UnitilsException;
 import org.unitils.dbmaintainer.dbsupport.DatabaseTask;
 import org.unitils.dbmaintainer.handler.StatementHandlerException;
@@ -24,21 +25,25 @@ import java.sql.SQLException;
 import java.util.*;
 
 /**
- * Implementation of {@link DBCleaner}. This implementation doesn't use any DBMS specific features, so it should work
- * for every database.
- * <p/>
- * // todo don't clear dbclearer itemstopreserve
+ * Implementation of {@link DBCleaner}. This implementation will delete all data from a database, except for the tables
+ * that are configured as tables to preserve. This includes the tables that are listed in the property
+ * {@link #PROPKEY_TABLESTOPRESERVE}, {@link #PROPKEY_DBCLEARER_ITEMSTOPRESERVE}. and the table that is configured as
+ * version table using the property {@link #PROPKEY_VERSION_TABLE_NAME}. 
+ *
+ * @author Filip Neven
  */
 public class DefaultDBCleaner extends DatabaseTask implements DBCleaner {
 
-    /* Property key for the database schema name */
-    public static final String PROPKEY_DATABASE_SCHEMANAME = "dataSource.schemaName";
+    private static final Logger logger = Logger.getLogger(DefaultDBCleaner.class);
 
     /* Property key for the tables that should not be cleaned */
     public static final String PROPKEY_TABLESTOPRESERVE = "dbMaintainer.cleanDb.tablesToPreserve";
 
+    /* Property key for the tables that should not be cleared (these tables should also not be cleaned */
+    public static final String PROPKEY_DBCLEARER_ITEMSTOPRESERVE = "dbMaintainer.clearDb.itemsToPreserve";
+
     /* The key of the property that specifies the name of the datase table in which the
-     * DB version is stored. This table should not be deleted */
+* DB version is stored. This table should not be deleted */
     public static final String PROPKEY_VERSION_TABLE_NAME = "dbMaintainer.dbVersionSource.tableName";
 
     /* The tables that should not be cleaned */
@@ -53,15 +58,18 @@ public class DefaultDBCleaner extends DatabaseTask implements DBCleaner {
         tablesToPreserve = new HashSet<String>();
         tablesToPreserve.add(configuration.getString(PROPKEY_VERSION_TABLE_NAME).toUpperCase());
         tablesToPreserve.addAll(toUpperCaseList(Arrays.asList(configuration.getStringArray(PROPKEY_TABLESTOPRESERVE))));
+        tablesToPreserve.addAll(toUpperCaseList(Arrays.asList(configuration.getStringArray(PROPKEY_DBCLEARER_ITEMSTOPRESERVE))));
     }
 
     /**
-     * Deletes all data from all tables in the database, except
+     * Deletes all data from the database, except for the tables that have been
+     * configured as <i>tablesToPreserve</i> , and the table in which the database version is stored
      *
      * @throws StatementHandlerException
      */
     public void cleanDatabase() throws StatementHandlerException {
         try {
+            logger.info("Cleaning database tables");
             Set<String> tables = dbSupport.getTableNames();
             tables.removeAll(tablesToPreserve);
             clearTables(tables);
@@ -76,9 +84,11 @@ public class DefaultDBCleaner extends DatabaseTask implements DBCleaner {
      * @param tableNames
      * @throws StatementHandlerException
      */
-    private void clearTables(Set<String> tableNames) throws StatementHandlerException {
+    private void clearTables(Set<String> tableNames) throws StatementHandlerException, SQLException {
         for (String tableName : tableNames) {
-            statementHandler.handle("delete from " + tableName);
+            if (dbSupport.getRecordCount(tableName) > 0) {
+                statementHandler.handle("delete from " + tableName);
+            }
         }
     }
